# Conflicts:
#	projects/Generate_Code/out/__default_SQL/IDefaultWCoinDAO.xml
#	projects/bit_trade/bittrade/bittrade-__default/src/main/resources/__default_SQL/IDefaultWCoinDAO.xml
#	projects/bit_trade/common-bittrade-service/src/main/resources/SQL/IWCoinDAO.xml

+++ projects/bit_trade/bittrade/bittrade-pojo/src/main/java/com/bittrade/pojo/vo/QueryAdvertVO.java
@@ -42,7 +42,7 @@ public class QueryAdvertVO implements Serializable {
 	 */
 	@NotNull(message = "AdvertType cannot be null")
 	@CheckEnumValue(enumClass = TAdvertInfoDTO.AdvertTypeEnum.class, enumMethod = "isValidAdvertType")
-	private Byte AdvertType;
+	private Integer AdvertType;
 
 	/**
 	 * 仅展示可交易
